+++ nextgisweb/basemap/nodepkg/layer-widget/LayerWidget.tsx
@@ -10,6 +10,7 @@ import type {
     EditorWidgetComponent,
     EditorWidgetProps,
 } from "@nextgisweb/resource/type";
+import { Basemap, MapComponent } from "@nextgisweb/webmap/preview-map";
 
 import type { LayerStore } from "./LayerStore";
 import { QMSSelect } from "./component/QMSSelect";
@@ -48,60 +49,95 @@ export const LayerWidget: EditorWidgetComponent<EditorWidgetProps<LayerStore>> =
         }, [qmsId, store.qms]);
 
         return (
-            <Area pad>
-                <Lot
-                    label={msgPickQms}
-                    help={() => (
-                        <>
-                            {msgPickQmsHelpMainPart}
-                            <a href={settings.qms_url} target="_blank">
-                                NextGIS QMS
-                            </a>
-                            . {msgPickQmsHelpTodoPart}.
-                        </>
-                    )}
-                >
-                    <QMSSelect
-                        value={qmsId}
-                        onChange={setQmsId}
-                        onService={(service) => {
-                            store.url.value = service.url;
-                            store.copyright_text.value = service.copyright_text;
-                            store.copyright_url.value = service.copyright_url;
-                            store.qms.value = JSON.stringify(service);
+            <>
+                <Area pad>
+                    <Lot
+                        label={msgPickQms}
+                        help={() => (
+                            <>
+                                {msgPickQmsHelpMainPart}
+                                <a href={settings.qms_url} target="_blank">
+                                    NextGIS QMS
+                                </a>
+                                . {msgPickQmsHelpTodoPart}.
+                            </>
+                        )}
+                    >
+                        <QMSSelect
+                            value={qmsId}
+                            onChange={setQmsId}
+                            onService={(service) => {
+                                store.url.value = service.url;
+                                store.copyright_text.value =
+                                    service.copyright_text;
+                                store.copyright_url.value =
+                                    service.copyright_url;
+                                store.qms.value = JSON.stringify(service);
+                            }}
+                        ></QMSSelect>
+                    </Lot>
+
+                    <LotMV
+                        help={disabled ? msgDisabled : undefined}
+                        value={store.url}
+                        component={InputValue}
+                        label={gettext("URL")}
+                        props={{
+                            disabled,
                         }}
-                    ></QMSSelect>
-                </Lot>
+                    />
 
-                <LotMV
-                    help={disabled ? msgDisabled : undefined}
-                    value={store.url}
-                    component={InputValue}
-                    label={gettext("URL")}
-                    props={{
-                        disabled,
-                    }}
-                />
-
-                <LotMV
-                    help={disabled ? msgDisabled : undefined}
-                    label={gettext("Copyright text")}
-                    value={store.copyright_text}
-                    component={InputValue}
-                    props={{
-                        disabled,
-                    }}
-                />
-                <LotMV
-                    help={disabled ? msgDisabled : undefined}
-                    label={gettext("Copyright URL")}
-                    value={store.copyright_url}
-                    component={InputValue}
-                    props={{
-                        disabled,
-                    }}
-                />
-            </Area>
+                    <LotMV
+                        help={disabled ? msgDisabled : undefined}
+                        label={gettext("Copyright text")}
+                        value={store.copyright_text}
+                        component={InputValue}
+                        props={{
+                            disabled,
+                        }}
+                    />
+                    <LotMV
+                        help={disabled ? msgDisabled : undefined}
+                        label={gettext("Copyright URL")}
+                        value={store.copyright_url}
+                        component={InputValue}
+                        props={{
+                            disabled,
+                        }}
+                    />
+                </Area>
+                {store.url.value ? (
+                    <MapComponent
+                        osm
+                        style={{
+                            margin: "1em",
+                            height: "400px",
+                            borderRadius: "4px",
+                            borderWidth: "1px",
+                            borderStyle: "solid",
+                            borderColor: "#d9d9d9",
+                        }}
+                        whenCreated={(adapter) => {
+                            const observer = new ResizeObserver(() => {
+                                adapter?.map.updateSize();
+                            });
+                            observer.observe(
+                                adapter?.map.getTargetElement() as Element
+                            );
+                        }}
+                    >
+                        <Basemap
+                            url={store.url.value}
+                            key={qmsId}
+                            attributions={
+                                store.copyright_url.value
+                                    ? `<a href="${store.copyright_url.value}" target="_blank">${store.copyright_text.value}</a>`
+                                    : store.copyright_text.value
+                            }
+                        />
+                    </MapComponent>
+                ) : null}
+            </>
         );
     });