+++ nextgisweb/basemap/view.py
@@ -1,3 +1,4 @@
+
 from nextgisweb.resource import Widget
 from nextgisweb.webmap import WebMap