+++ nextgisweb/feature_layer/nodepkg/type/FeatureExtent.ts
@@ -1,3 +1,5 @@
+/** @import NgwExtent from "@nextgisweb/feature-layer/type/api" */
+/** @deprecated Use @type {NgwExtent}} from "@nextgisweb/feature-layer/type/api" */
 export interface NgwExtent {
     minLon: number;
     minLat: number;