+++ nextgisweb/layer_preview/nodepkg/preview-layer/index.ts
@@ -1,0 +1,5 @@
+/** @entrypoint */
+
+import { PreviewLayer } from "./PreviewLayer";
+
+export default PreviewLayer;