+++ nextgisweb/resource/nodepkg/children-section/util/isPreviewAction.ts
@@ -1,0 +1,6 @@
+import type { ChildrenResourceAction } from "../type";
+
+export function isPreviewAction(action: ChildrenResourceAction) {
+    const { key } = action;
+    return Array.isArray(key) && key[1] === "preview";
+}