+++ nextgisweb/webmap/nodepkg/preview-map/index.ts
@@ -1,0 +1,7 @@
+import { Basemap } from "./component/Basemap";
+import { MapComponent } from "./component/MapComponent";
+import { NGWLayer } from "./component/NGWLayer";
+import type { LayerType } from "./component/hook/useNGWLayer";
+
+export { Basemap, NGWLayer, MapComponent };
+export type { LayerType };